Split HistoryDB modules
