+++ apps/server/vibesensor/adapters/persistence/history_db/_sample_io.py
@@ -1,0 +1,105 @@
+"""Sample streaming helpers for ``HistoryDB``."""
+
+from __future__ import annotations
+
+import json
+import logging
+import sqlite3
+from collections.abc import Iterator
+from contextlib import AbstractContextManager
+
+from vibesensor.adapters.persistence.history_db._samples import (
+    ALLOWED_SAMPLE_TABLES,
+    V2_SELECT_SQL_COLS,
+    v2_row_to_dict,
+)
+from vibesensor.shared.types.json_types import JsonObject
+
+LOGGER = logging.getLogger(__name__)
+
+
+class _HistoryDBSampleIOMixin:
+    def _cursor(self, *, commit: bool = True) -> AbstractContextManager[sqlite3.Cursor]:
+        raise NotImplementedError
+
+    def get_run_samples(self, run_id: str) -> list[JsonObject]:
+        rows: list[JsonObject] = []
+        for batch in self.iter_run_samples(run_id):
+            rows.extend(batch)
+        return rows
+
+    def iter_run_samples(
+        self,
+        run_id: str,
+        batch_size: int = 1000,
+        offset: int = 0,
+    ) -> Iterator[list[JsonObject]]:
+        if offset < 0:
+            raise ValueError(f"iter_run_samples: offset must be >= 0, got {offset}")
+        yield from self._iter_v2_samples(run_id, batch_size, offset)
+
+    def _resolve_keyset_offset(
+        self,
+        table: str,
+        run_id: str,
+        offset: int,
+    ) -> int | None:
+        if table not in ALLOWED_SAMPLE_TABLES:
+            raise ValueError(
+                f"_resolve_keyset_offset: invalid table name {table!r}; "
+                f"must be one of {sorted(ALLOWED_SAMPLE_TABLES)}",
+            )
+        with self._cursor(commit=False) as cur:
+            cur.execute(
+                f"SELECT id FROM {table} WHERE run_id = ? ORDER BY id LIMIT 1 OFFSET ?",
+                (run_id, offset - 1),
+            )
+            row = cur.fetchone()
+        return int(row[0]) if row else None
+
+    def _iter_v2_samples(
+        self,
+        run_id: str,
+        batch_size: int = 1000,
+        offset: int = 0,
+    ) -> Iterator[list[JsonObject]]:
+        size = max(1, batch_size)
+        last_id: int | None = None
+        if offset > 0:
+            last_id = self._resolve_keyset_offset("samples_v2", run_id, offset)
+            if last_id is None:
+                return
+        total_skipped = 0
+        while True:
+            with self._cursor(commit=False) as cur:
+                if last_id is None:
+                    cur.execute(
+                        f"SELECT {V2_SELECT_SQL_COLS} FROM samples_v2"
+                        " WHERE run_id = ? ORDER BY id LIMIT ?",
+                        (run_id, size),
+                    )
+                else:
+                    cur.execute(
+                        f"SELECT {V2_SELECT_SQL_COLS} FROM samples_v2"
+                        " WHERE run_id = ? AND id > ? ORDER BY id LIMIT ?",
+                        (run_id, last_id, size),
+                    )
+                batch_rows = cur.fetchall()
+            if not batch_rows:
+                if total_skipped:
+                    LOGGER.warning(
+                        "run_id=%s: skipped %d corrupt v2 sample row(s) in total",
+                        run_id,
+                        total_skipped,
+                    )
+                return
+            last_id = batch_rows[-1][0]
+            parsed_batch: list[JsonObject] = []
+            for row in batch_rows:
+                try:
+                    parsed_batch.append(v2_row_to_dict(row))
+                except (json.JSONDecodeError, KeyError, ValueError, TypeError):
+                    total_skipped += 1
+                    LOGGER.warning("Skipping corrupt v2 sample row id=%s", row[0], exc_info=True)
+            if parsed_batch:
+                yield parsed_batch
